refactor: Create a new trait for using the alerts
I created a new Trait for using the alers outside the actions.
